Address PR feedback: integer LastInsertRowid, optional query return, unused var fixes
- Change LastInsertRowid from float to integer type in Python (int|None),
  Go (*int64), and Rust (Option<i64>). Adapters convert to wire float types.
- Make sqlite_query return optional in Python (| None) and Rust (Option<>),
  matching Node.js and .NET behavior for exec-type queries.
- Fix unused msg variable in Python E2E test.
- Remove unused import in Rust E2E test.

+++ go/internal/e2e/session_fs_sqlite_e2e_test.go
@@ -214,7 +214,7 @@ func (p *inMemorySqliteProvider) SqliteQuery(queryType rpc.SessionFsSqliteQueryT
 	case rpc.SessionFsSqliteQueryTypeExec:
 		return &copilot.SessionFsSqliteQueryResult{Columns: []string{}, Rows: []map[string]any{}}, nil
 	case rpc.SessionFsSqliteQueryTypeRun:
-		lastID := float64(1)
+		lastID := int64(1)
 		return &copilot.SessionFsSqliteQueryResult{
 			Columns:         []string{},
 			Rows:            []map[string]any{},
